feat: add a keyboard shortcut to run code
